Migrate flightActions into main file
